% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpress}
\alias{canvasXpress}
\title{Creates a CanvasXpress object
Refer to the CanvasXpress original documentation for more details: \code{\link{http://canvasxpress.org/html/documentation.html}}}
\usage{
canvasXpress(data = NULL, decorData = NULL, smpAnnot = NULL,
  varAnnot = NULL, nodeData = NULL, edgeData = NULL, vennData = NULL,
  vennLegend = NULL, genomeData = NULL, graphType = "Scatter2D",
  events = NULL, afterRender = NULL, width = 600, height = 400,
  pretty = FALSE, digits = 4, ...)
}
\description{
Creates a CanvasXpress object
Refer to the CanvasXpress original documentation for more details: \code{\link{http://canvasxpress.org/html/documentation.html}}
}

