% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvasXpress.R
\name{canvasXpressHTML}
\alias{canvasXpressHTML}
\title{Generates an HTML page given an canvasXpress object}
\usage{
canvasXpressHTML(cx, id = "canvas", width = "540", height = "640")
}
\arguments{
\item{cx}{CanvasXpress Object from \code{\link{canvasXpress}}}

\item{id}{HTML5 canvas ID}

\item{width}{HTML5 canvas width}

\item{height}{HTML5 canvas height}
}
\description{
Generates an HTML page given an canvasXpress object
}

