% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cXplot.R
\name{cXboxplot}
\alias{cXboxplot}
\title{Writes HTML page with interactive boxplots for each row of an input matrix}
\usage{
cXboxplot(mat, groups)
}
\arguments{
\item{mat}{Matrix}

\item{groups}{Factor grouping. Should match columns of matrix.}
}
\description{
Writes HTML page with interactive boxplots for each row of an input matrix
}
\examples{
data(mtcars)
mat <- t(mtcars)
groups <- mtcars$cyl
write(cXboxplot(mat, groups), file="boxplot.html")

}

