% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cXplot.R
\name{cXheatmap}
\alias{cXheatmap}
\title{Writes HTML page with an interactive heatmap for input matrix}
\usage{
cXheatmap(mat)
}
\arguments{
\item{mat}{Matrix}
}
\description{
Writes HTML page with an interactive heatmap for input matrix
}
\examples{
data(mtcars)
mtcars <- scale(mtcars)
mtcars <- t(scale(t(mtcars)))
write(cXheatmap(mtcars), file="heatmap.html")

}

