% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cXplot.R
\name{cXscatter}
\alias{cXscatter}
\title{Writes HTML page with interactive scatterplot given a set of X and Y values}
\usage{
cXscatter(df, groups)
}
\arguments{
\item{df}{2xN dataframe or matrix where first column is X values and second is Y values}

\item{groups}{Factor groupings for each value}
}
\description{
Writes HTML page with interactive scatterplot given a set of X and Y values
}
\examples{
data(mtcars)
basePCA <- prcomp(mtcars)
groups <- paste0("cyl:", mtcars$cyl)
df <- basePCA$x[,1:2]
write(cXscatter(df, groups), file="scatter.html")

}

