% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{writeDataString}
\alias{writeDataString}
\title{Converts matrix to Javascript array data structure}
\usage{
writeDataString(mat)
}
\arguments{
\item{mat}{Matrix}
}
\description{
Converts matrix to Javascript array data structure
}
\examples{
data(mtcars)
writeDataString(head(mtcars))

}

