% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cXplot.R
\name{cXboxplot}
\alias{cXboxplot}
\title{Writes HTML page with interactive boxplots for each row of an input matrix}
\usage{
cXboxplot(mat, groups)
}
\arguments{
\item{mat}{Matrix}
}
\description{
Writes HTML page with interactive boxplots for each row of an input matrix
}
\examples{
data(mtcars)
write(cXboxplot(t(mtcars[, c("mpg", "hp", "wt")]), mtcars$cyl), file="boxplot.html")

}

